backend/requirements.txt
fastapi
uvicorn
sqlalchemy
google-generativeai
youtube-search-python
googlemaps
python-jose[cryptography]
pydantic
pydantic-settings
